requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.14
pydantic==2.5.0
requests==2.31.0
streamlit==1.28.0
python-dotenv==1.0.0
google-generativeai==0.3.2
pandas==2.1.3
python-multipart==0.0.6
pymysql==1.1.0
cryptography==41.0.7
sqlalchemy==2.0.23
groq==0.9.0